Rag_Agent/requirements.txt
flask==2.3.6
python-dotenv==1.1.1
qdrant-client==1.15.1
langchain==0.3.76
langchain-community==0.3.29
sentence-transformers==2.2.2
torch==2.8.0  # for HuggingFace embeddings
langchain-huggingface==0.3.1
